+++ service/firehose/doc.go
@@ -7,7 +7,8 @@
 //
 // Amazon Data Firehose is a fully managed service that delivers real-time streaming
 // data to destinations such as Amazon Simple Storage Service (Amazon S3), Amazon
-// OpenSearch Service, Amazon Redshift, Splunk, and various other supportd destinations.
+// OpenSearch Service, Amazon Redshift, Splunk, and various other supported
+// destinations.
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/firehose-2015-08-04 for more information on this service.
 //