+++ service/sqs/api.go
@@ -2322,12 +2322,15 @@ func (c *SQS) SendMessageRequest(input *SendMessageInput) (req *request.Request,
 // Delivers a message to the specified queue.
 //
 // A message can include only XML, JSON, and unformatted text. The following
-// Unicode characters are allowed:
+// Unicode characters are allowed. For more information, see the W3C specification
+// for characters (http://www.w3.org/TR/REC-xml/#charsets).
 //
 // #x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF
 //
-// Any characters not included in this list will be rejected. For more information,
-// see the W3C specification for characters (http://www.w3.org/TR/REC-xml/#charsets).
+// Amazon SQS does not throw an exception or completely reject the message if
+// it contains invalid characters. Instead, it replaces those invalid characters
+// with U+FFFD before storing the message in the queue, as long as the message
+// body contains at least one valid character.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -2478,12 +2481,15 @@ func (c *SQS) SendMessageBatchRequest(input *SendMessageBatchInput) (req *reques
 // both 256 KiB (262,144 bytes).
 //
 // A message can include only XML, JSON, and unformatted text. The following
-// Unicode characters are allowed:
+// Unicode characters are allowed. For more information, see the W3C specification
+// for characters (http://www.w3.org/TR/REC-xml/#charsets).
 //
 // #x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF
 //
-// Any characters not included in this list will be rejected. For more information,
-// see the W3C specification for characters (http://www.w3.org/TR/REC-xml/#charsets).
+// Amazon SQS does not throw an exception or completely reject the message if
+// it contains invalid characters. Instead, it replaces those invalid characters
+// with U+FFFD before storing the message in the queue, as long as the message
+// body contains at least one valid character.
 //
 // If you don't specify the DelaySeconds parameter for an entry, Amazon SQS
 // uses the default value for the queue.
@@ -8294,12 +8300,15 @@ type SendMessageInput struct {
 	// is 256 KiB.
 	//
 	// A message can include only XML, JSON, and unformatted text. The following
-	// Unicode characters are allowed:
+	// Unicode characters are allowed. For more information, see the W3C specification
+	// for characters (http://www.w3.org/TR/REC-xml/#charsets).
 	//
 	// #x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF
 	//
-	// Any characters not included in this list will be rejected. For more information,
-	// see the W3C specification for characters (http://www.w3.org/TR/REC-xml/#charsets).
+	// Amazon SQS does not throw an exception or completely reject the message if
+	// it contains invalid characters. Instead, it replaces those invalid characters
+	// with U+FFFD before storing the message in the queue, as long as the message
+	// body contains at least one valid character.
 	//
 	// MessageBody is a required field
 	MessageBody *string `type:"string" required:"true"`